+++ gain/mp_gain_year_count_planted_forest.py
@@ -66,10 +66,10 @@ def mp_gain_year_count_planted_forest(sensit_type, tile_id_list, run_date = None
 
     # Creates gain year count tiles using only pixels that had only loss
     # count/3 maxes out at about 300 GB
-    pool = multiprocessing.Pool(cn.count/2)
+    pool = multiprocessing.Pool(int(cn.count/2))
     pool.map(gain_year_count_planted_forest.create_gain_year_count_loss_only, tile_id_list)
 
-    pool = multiprocessing.Pool(cn.count/2)
+    pool = multiprocessing.Pool(int(cn.count/2))
     if sensit_type == 'maxgain':
         # Creates gain year count tiles using only pixels that had only gain
         pool.map(gain_year_count_planted_forest.create_gain_year_count_gain_only_maxgain, tile_id_list)
@@ -79,11 +79,11 @@ def mp_gain_year_count_planted_forest(sensit_type, tile_id_list, run_date = None
 
     # Creates gain year count tiles using only pixels that had neither loss nor gain pixels
     # count/3 maxes out at 260 GB
-    pool = multiprocessing.Pool(cn.count/2)
+    pool = multiprocessing.Pool(int(cn.count/2))
     pool.map(gain_year_count_planted_forest.create_gain_year_count_no_change, tile_id_list)
 
     # count/3 maxes out at about 230 GB
-    pool = multiprocessing.Pool(cn.count/2)
+    pool = multiprocessing.Pool(int(cn.count/2))
     if sensit_type == 'maxgain':
         # Creates gain year count tiles using only pixels that had only gain
         pool.map(gain_year_count_planted_forest.create_gain_year_count_loss_and_gain_maxgain, tile_id_list)
@@ -96,7 +96,7 @@ def mp_gain_year_count_planted_forest(sensit_type, tile_id_list, run_date = None
 
     # Merges the four above gain year count tiles for each Hansen tile into a single output tile
     # Count/6 maxes out at about 220 GB (doesn't increase all the way through GDAL process for some reason)
-    pool = multiprocessing.Pool(cn.count/6)
+    pool = multiprocessing.Pool(int(cn.count/6))
     pool.map(partial(gain_year_count_planted_forest.create_gain_year_count_merge, pattern=pattern), tile_id_list)
     pool.close()
     pool.join()
